cleaned up piano files
